Revert improved input size estimates
This reverts commit decf974cb6046bc677bea7cd9c8a5a54c2af59ec.

The change allows us to use a typescript version of this package on
IMS and www, which are not yet compatible with the new and improved
input sizes.

The tags might be a little confusing now, so let me explain:

* Tag 0.1.0: initial package version, not used anywhere anymore.
* Tag 0.2.0: some smaller improvements, no breaking changes. Not used
  anywhere
* Tag 0.2.9: Support for dynamic output sizes, but no change in input
  sizes. This version is used on www and IMS (including subpackages).
  However it used in a way so that fixed-sizes outputs are computed,
  to avoid compatibility problems.
* Tag 0.3.0: Tweaked input sizes.
* Tag 0.3.1: bumped version in `package.json` because I missed it in
  0.3.0.
* Tag 0.4.0: typescript port

To keep with the theme that tag `0.2.x` has the old input sizes, I
want to tag this version `0.2.10`. It is basically the same as `0.2.9`
but ported to typescript.