Declare used type data explicitly
tl;dr: Previously, the package could not be used as a dependency in
`www/`, now it can.

Long boring explanation:

When `tsc` compiles a package, it traverses the `node_modules` in
parent directories. In the case of this project being imported in
`platform/www`, this seems to include broken type definitions in
`www/node_modules/@types/` which were clearly unrelated to this
package. By explicitly declaring the type definitions in
`tsconfig.json`, we only look at typedefs we actually care about when
invoking `tsc` and avoid loading broken ones.

Issue: BG-9658