feat: add partial support for p2tr script types
This adds partial support for the `p2tr` script type to `Codes` and
`Dimensions`.

The remaining work needed for full support is calculation of input sizes
for p2tr multisig script path spends, which can vary depending on which
of the three scripts are used in a scenario where a 2 of 3 multisignature
scheme is split into a tap tree with three 2 of 2 scripts for each
combination. For testing purposes this will require signing/spending p2tr
support which will be available in a future version of `@BitGo/utxo-lib`.

Ticket: BG-37329