build: update & revise dev dependencies
This makes several changes to the dev dependencies used for testing
and typescript compilation.

- The outdated `bitgo-utxo-lib` is replaced with `@bitgo/utxo-lib` which
includes p2tr output script support.

- The `hdkey` dependency is replaced with `bip32`. `hdkey` uses a
version of the `secp256k1` package which is not compatible with more
recent versions of NodeJS. `bip32` is already included with the
`@bitgo/utxo-lib` dependency.

- `@node/types` is updated to v14.17.5. The previous 10.x version is
incompatible with `@bitgo/utxo-lib` which requires `Object.fromEntries`
only available in NodeJS versions 12.4 and higher.

- `typescript` is updated to the latest version.

Drone build steps using versions of node lower than 12.x are removed and
replaced with newer versions.

All tests pass with the updates from this commit.